+++ src/Localisation/Currency.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace OzdemirBurak\SkyScanner\Localisation;
+
+use OzdemirBurak\SkyScanner\LocalisationService;
+
+class Currency extends LocalisationService
+{
+    /**
+     * @var string
+     */
+    protected $property = 'Currencies';
+
+    /**
+     * @return string
+     */
+    public function getUri()
+    {
+        return 'reference/v1.0/currencies';
+    }
+}